+++ src/renderer/components/page.tsx
@@ -11,23 +11,25 @@ import { useDrawer } from '../hooks/use-drawer'
 
 interface Props {
   className?: string
+  mainClassName?: string
 }
 
 export function Page({
   children,
   className,
+  mainClassName,
 }: React.PropsWithChildren<Props>): JSX.Element {
   const [isDrawerExpand] = useDrawer()
 
   return (
     <div
       className={cx(
-        'page h-screen w-screen p-6 overflow-overlay bg-light-600 dark:bg-black-800',
+        'page h-screen w-screen p-6 overflow-overlay bg-light-600 dark:bg-black-800 transition-all duration-300',
         className,
         isDrawerExpand ? 'pl-48' : 'pl-14',
       )}
     >
-      <main>{children}</main>
+      <main className={mainClassName}>{children}</main>
     </div>
   )
 }